Add a call to get the screen size
Also replaced hardcoded LEM-1802 sized with defines LEM_WID and
LEM_HGT.

Since there may be different kinds of screens (Tech Compliant,
for example, wants to add a whole series of little ones), the
BBOS API should be capable of reporting the screen size.

Yet to be determined is whether support for other screens should
be added by adding code to BBOS, or by maintaining
hardware-specific forks.

Note that bbos.simple.asm and rom.bin are not updated with this
change. How are those generated?