Added script to generate fastcall BBOS wrappers
BBOS calls are what are traditionally known as stdcall. This means
that all arguments are pushed to the stack.
The fastcall calling convention uses registers for arguments instead
of the stack. As some people may prefer to use fastcall over stdcall,
this script generates wrappers which converts fastcalls to BBOS stdcalls