Reverse documented return order of height and width for screen size
The code for getting the screen size is:

```
.getsize:
    SET [Z+0], LEM_WID
    SET [Z+1], LEM_HGT
    SET PC, POP
```

This clearly returns the height in the argument that was pushed first
(and will be popped last) and the width in the argument that was pushed
last (and will be popped first).

The documentation specified the other order. This PR changes the
documentation to match the code.