LEM/PIXIE: Only set the VRAM once on activation.
While it may be safer to always set the VRAM every time the screen
is written to incase someone else has changed it, the new
'set active screen' creates a new middle ground. If the user has
modified a displays state, it can be reset by 'reactivating' the
display with 'set active screen'.